third_party/blink/web_tests/fast/url/anchor-expected.txt

Test URLs that have an anchor.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS canonicalize('http://www.example.com/#hello, world') is 'http://www.example.com/#hello,%20world'
PASS canonicalize('http://www.example.com/#Â©') is 'http://www.example.com/#%C3%82%C2%A9'
PASS canonicalize('http://www.example.com/#𐌀ss') is 'http://www.example.com/#%F0%90%8C%80ss'
PASS canonicalize('http://www.example.com/#%41%a') is 'http://www.example.com/#%41%a'
PASS canonicalize('http://www.example.com/#\ud800\u597d') is 'http://www.example.com/#%EF%BF%BD%E5%A5%BD'
PASS canonicalize('http://www.example.com/#a\uFDD0') is 'http://www.example.com/#a%EF%B7%90'
PASS canonicalize('http://www.example.com/#asdf#qwer') is 'http://www.example.com/#asdf#qwer'
PASS canonicalize('http://www.example.com/##asdf') is 'http://www.example.com/##asdf'
PASS canonicalize('http://www.example.com/#a\nb\rc\td') is 'http://www.example.com/#abcd'
PASS successfullyParsed is true

TEST COMPLETE
